requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.3
djangorestframework==3.12.2
jwt==1.2.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
